Updated map interface panels
