+++ packages/map-interface/src/expansion-panel/main.module.sass
@@ -1,24 +1,28 @@
 
 .panel-subhead
-  padding: 0.2em var(--box-horizontal-padding)
-  border-top: 1px solid var(--panel-rule-color)
+  padding: 0 1rem
+  border: 1px solid var(--panel-rule-color)
   border-bottom: 1px solid var(--panel-rule-color)
   background-color: var(--accent-color)
   display: flex
   flex-direction: row
   align-items: center
   z-index: 1
-  gap: var(--box-horizontal-padding)
+  gap: var(--panel-header-gap, 0.5rem)
   top: -1px
   position: sticky
+  min-height: var(--panel-header-height, 40px)
   h1, h2, h3, h4
     font-family: Montserrat,sans-serif
     font-weight: 700
-    margin: 0.2em 0
+    margin: 0
   h4
     font-weight: 600
   .title
     flex-grow: 1
+
+  .expansion-panel-icon
+    margin-right: -0.35em
 
 // :global(.bp5-dark) .panel-subhead
 //   margin 0 1px
@@ -48,14 +52,14 @@
   .expansion-panel-header
     background-color: var(--accent-secondary-color)
     cursor: pointer
+    padding: 0 1rem
+
     &:hover
       background-color: var(--accent-secondary-hover-color)
     h2, h3, h4
       font-weight: 500
       border-bottom: 1px solid var(--tertiary-border-color)
       border-top: 1px solid var(--tertiary-border-color)
-      margin-top: -1px
-      padding: 5px 1em 5px
       align-items: center
 
 .expansion-summary-title-help
@@ -71,12 +75,18 @@
     transform: translate(0,3px)
 
 .expansion-children
-  padding: 5px 1em 10px
+  padding: 0 1rem
   .expansion-panel
-    margin-left: -1em
-    margin-right: -1em
+    margin-left: -1rem
+    margin-right: -1rem
     &:first-child
-      margin-top: -5px
+      margin-top: 0
+  &>:first-child
+    margin-top: 1rem
+  &>:last-child
+    margin-bottom: 1rem
+
+
 
 .expansion-panel-subtext
   font-size: 85%
